Add instructions to the HTML webpage
I knew I had to do this initially, but for some reason I just neglected
to do so. I have added a section within the HTML page, which are the
instructions to the user in how to use the app in terms of colouring in
and erasing any coloured-in pixels.