fix: Add a removeGrid function to the JS file
After a good deal of testing, I have wrote a function that will clear or
wipe away an existing grid should the user choose to create another
grid while there is an existing grid. This function fixes the issue in
the earlier build where new grids would be stacked beneath the existing
grids, which is obviously a serious flaw.